+++ ksa_compliance/hooks.py
@@ -135,6 +135,7 @@ doc_events = {
     "Sales Invoice": {
         "on_submit": "ksa_compliance.standard_doctypes.sales_invoice.create_sales_invoice_additional_fields_doctype",
         "before_save": "ksa_compliance.standard_doctypes.sales_invoice.calculate_tax_amount",
+        "validate": "ksa_compliance.standard_doctypes.sales_invoice.validate_tax_category",
         "before_cancel": "ksa_compliance.standard_doctypes.sales_invoice.prevent_cancellation_of_sales_invoice"
     },
 }